backend/requirements.txt
# ── Web Framework ─────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9          # multipart form / file uploads

# ── Database ──────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0                  # async PostgreSQL driver
alembic==1.13.1
psycopg2-binary==2.9.9           # Alembic sync driver

# ── Authentication ────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Validation & Config ───────────────────────
pydantic==2.7.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# ── HTTP Client (ML service calls) ────────────
httpx==0.27.0

# ── Caching & Task Queue ──────────────────────
redis==5.0.4
celery==5.4.0
celery[redis]==5.4.0

# ── Rate Limiting ─────────────────────────────
slowapi==0.1.9

# ── Data Processing ───────────────────────────
pandas==2.2.2
openpyxl==3.1.2

# ── Structured Logging ────────────────────────
structlog==24.1.0

# ── Testing ───────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# ── Linting / Type Checking ───────────────────
ruff==0.4.4
mypy==1.10.0
